requirements.txt
requests
python-dotenv
pyyaml
beautifulsoup4
playwright